Load 3.3.175 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@406 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/mapping.c
@@ -3966,7 +3966,7 @@ f_mkmapping (svalue_t *sp, int num_arg)
 
     if (max_mapping_size && length * num_arg > max_mapping_size)
         error("Illegal mapping size: %ld elements (%ld x %ld)\n"
-             , length * num_arg, length, num_arg);
+             , length * num_arg, length, (long)num_arg);
 
     /* Allocate the mapping */
     num_values = num_arg - 1;
